Move Grafana datasource/dashboard functions and state to ApiService
